Move event logging from separate component to DevToolsPanel
