+++ src/App.tsx
@@ -2,7 +2,6 @@
 import pkg from '../package.json'
 import { Board } from './components/Board'
 import { DevToolsPanel } from './components/DevToolsPanel'
-import { EventLogger } from './components/EventLogger'
 // import { ImportanceSlider } from './components/ImportanceSlider'
 import { IntroScreen } from './components/IntroScreen'
 import PWAInstallPrompt from './components/PWAInstallPrompt'
@@ -27,7 +26,6 @@ const AppContent: React.FC = () => {
           <DecksProvider>
             <AppMain />
             <DevToolsPanel />
-            <EventLogger />
             {/* <ImportanceSlider /> */}
           </DecksProvider>
         </GamesProvider>
